+++ NEWS.md
@@ -1,7 +1,8 @@
 # TreeTools 1.10.0.9001 (development) #
 
 - `DescendantTips()` complements `DescendantEdges()`.
-- Document breaking of external edges in `AddTip()`.
+- `AddTip(edgeLength = NULL)` defaults to `lengthBelow`. This will become the
+  default in a future release.
 
 
 # TreeTools 1.10.0 (2023-08-18) #